change parameters type

+++ SizeJudgmentformula.py
@@ -4,9 +4,9 @@ class Size:
     """
 
     def __init__(self, single_dict):
-        self.income = int(single_dict['income'])
-        self.num_people = int(single_dict['num_people'])
-        self.total_assets = int(single_dict['total_assets'])
+        self.income = float(single_dict['income'])
+        self.num_people = float(single_dict['num_people'])
+        self.total_assets = float(single_dict['total_assets'])
 
     def single_income_jg(self, _max, _2nd, _min):
         if self.income >= _max:
